pyproject.toml
[tool.poetry]
name = "bingo"
version = "0.1.0"
description = ""
authors = ["Alynne Ferreira <alynneferreiras@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^9.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
